requirements.txt
torch==1.11.0
nibabel==3.2.2
MedPy==0.4.0
opencv-python==4.5.5.64
einops==0.4.1
numpy==1.17.5
sklearn
matplotlib==3.1.2
jupyter==1.0.0
ipympl==0.8.8
h5py==3.1.0
pyyaml==6.0
pydicom==2.3.0
torchvision==0.12.0
rasterio==1.2.10
pandas==1.3.5
tqdm==4.64.0
importlib_metadata==4.11.4